Enemy_03_BP // Shooting mechanic
New:

Enemy 03 now stops at random moments and shoots (This doesent happen if there it a block missing next to the enemy.

When he is traped in the corner the enemy will shoot in the direction he is faceing.

Bug

Enemy 03 floats in the air sometimes when there is a block underneath smashed.